+++ clang/docs/ReleaseNotes.rst
@@ -89,6 +89,8 @@ C++2c Feature Support
 
 Resolutions to C++ Defect Reports
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Implemented `DR2137 <https://wg21.link/CWG2137>`_ which allows
+  list-initialization from objects of the same type.
 
 C Language Changes
 ------------------